fixes and interface file creation
Deleted all refrences to the editor state as well.
