+++ src/Sprites/index.ts
@@ -1,4 +1,5 @@
 export * from "./sprite";
+export * from "./ISprites";
 export * from "./spriteManager";
 export * from "./spriteMap";
 export * from "./spritePackedManager";
